configuração de persistência com PostgreSQL, EF Core e migrations

+++ src/RaizesNordeste.Domain/Entities/ProdutoItemInventario.cs
@@ -9,7 +9,7 @@ namespace RaizesNordeste.Domain.Entities
         public Produto Produto { get; set; } = null!;
         public long ItemInventarioId { get; set; }
         public ItemInventario ItemInventario { get; set; } = null!;
-        public decimal QuantidadeConsumo { get; set; }
+        public int QuantidadeConsumo { get; set; }
         public long CriadoPor {  get; set; }
         public Usuario UsuarioCriador { get; set; } = null!;
         public DateTime DataCriacao { get; set; }
